Prototype of CLI and VirtualView
Plan A (current plan) : we use JSON to transmit the virtual view.
Pro : just a little modification then we get the game saving feature.
Con : anyway it's a new thing for us, we need to learn it first, but seems not too hard.
Plan B : we create two the same Virtual View class on both client and server.
Pro : easy to realize
Con : less extensibility